refactor: run browserify and mocha directly
Stop copying them in a flat staging area. To avoid
long and messy relative filenames, use a package
that implements module aliases for node. That's ok
for mocha, but does not work as is for mocha. So I
implementeed a custom browserify transformer which
handles these aliases.
Also bumped yarn to 45.1. Cannot make workspaces
yet, as dependencies would be hoisted and grunt
would be lost bcause it could not find its plugins
locally.