refactor: made app a monorepo
Turned out I just needed to udate grunt to handle mono-repos.
Maybe trying to remove it wasn't necessary. Having a monorepo
causes an bbissue with module-alias though, I must redefine
aliases in the root package.json. They are duplicated, this
is annoying. For now though, having a monorepo opens
up the possibility of running tests in a neat CI pipeline.