fix: add cleanup in log-game-task test
Ran the tests a few times and now they complete
right afteer they are ran, they don't hang. What
I did was adding an afterEach in the LogGameTask
test to stop the game it started to test its
monitoring, since the random error seemed to be
caused by a game running wild and sometimes
ending badly.